smart-resume-coverletter/requirements.txt
streamlit>=1.36,<2
python-dotenv>=1.0,<2
openai>=1.35,<2
python-docx>=1.1,<2
pdfplumber>=0.11,<2
tenacity>=8.2,<9